+++ tidy3d/web/webapi.py
@@ -4,7 +4,11 @@ import os
 import time
 from datetime import datetime, timedelta
 from typing import List, Dict, Callable
+from functools import wraps
+
 from requests import HTTPError
+from requests.exceptions import ConnectionError as ConnErr
+from urllib3.exceptions import NewConnectionError
 
 import pytz
 from rich.console import Console
@@ -31,7 +35,39 @@ TOTAL_DOTS = 3
 # file names when uploading to S3
 SIM_FILE_JSON = "simulation.json"
 
+# number of seconds to keep re-trying connection before erroring
+CONNECTION_RETRY_TIME = 30
 
+
+def wait_for_connection(wait_time_sec: float = CONNECTION_RETRY_TIME):
+    """Causes function to ignore connection errors and retry for ``wait_time_sec`` secs."""
+
+    def decorator(web_fn):
+        """Decorator returned by @wait_for_connection()"""
+
+        @wraps(web_fn)
+        def web_fn_wrapped(*args, **kwargs):
+            """Function to return including connection waiting."""
+            time_start = time.time()
+            warned_previously = False
+
+            while (time.time() - time_start) < wait_time_sec:
+                try:
+                    return web_fn(*args, **kwargs)
+                except (ConnErr, ConnectionError, NewConnectionError):
+                    if not warned_previously:
+                        log.warning(f"No connection: Retrying for {wait_time_sec} seconds.")
+                        warned_previously = True
+                    time.sleep(REFRESH_TIME)
+
+            raise WebError("No internet connection: giving up on connection waiting.")
+
+        return web_fn_wrapped
+
+    return decorator
+
+
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def run(  # pylint:disable=too-many-arguments
     simulation: Simulation,
     task_name: str,
@@ -95,6 +131,7 @@ def run(  # pylint:disable=too-many-arguments
     )
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def upload(  # pylint:disable=too-many-locals,too-many-arguments
     simulation: Simulation,
     task_name: str,
@@ -157,6 +194,7 @@ def upload(  # pylint:disable=too-many-locals,too-many-arguments
     return task.task_id
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def get_info(task_id: TaskId) -> TaskInfo:
     """Return information about a task.
 
@@ -176,6 +214,7 @@ def get_info(task_id: TaskId) -> TaskInfo:
     return TaskInfo(**{"taskId": task.task_id, **task.dict()})
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def start(
     task_id: TaskId,
     solver_version: str = None,
@@ -203,6 +242,7 @@ def start(
     task.submit(solver_version=solver_version, worker_group=worker_group)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def get_run_info(task_id: TaskId):
     """Gets the % done and field_decay for a running task.
 
@@ -353,6 +393,7 @@ def monitor(task_id: TaskId, verbose: bool = True) -> None:
             time.sleep(REFRESH_TIME)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def download(
     task_id: TaskId,
     path: str = "simulation_data.hdf5",
@@ -377,6 +418,7 @@ def download(
     task.get_simulation_hdf5(path, verbose=verbose, progress_callback=progress_callback)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def download_json(
     task_id: TaskId,
     path: str = SIM_FILE_JSON,
@@ -402,6 +444,7 @@ def download_json(
     task.get_simulation_json(path, verbose=verbose, progress_callback=progress_callback)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def download_hdf5(
     task_id: TaskId,
     path: str = SIM_FILE_HDF5,
@@ -427,6 +470,7 @@ def download_hdf5(
     task.get_simulation_hdf5(path, verbose=verbose, progress_callback=progress_callback)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def load_simulation(
     task_id: TaskId,
     path: str = SIM_FILE_JSON,
@@ -458,6 +502,7 @@ def load_simulation(
     return Simulation.from_file(path)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def download_log(
     task_id: TaskId,
     path: str = "tidy3d.log",
@@ -485,6 +530,7 @@ def download_log(
     task.get_log(path, verbose=verbose, progress_callback=progress_callback)
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def load(
     task_id: TaskId,
     path: str = "simulation_data.hdf5",
@@ -534,6 +580,7 @@ def load(
     return sim_data
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def delete(task_id: TaskId) -> TaskInfo:
     """Delete server-side data associated with task.
 
@@ -554,6 +601,7 @@ def delete(task_id: TaskId) -> TaskInfo:
     return TaskInfo(**{"taskId": task.task_id, **task.dict()})
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def delete_old(
     days_old: int = 100,
     folder: str = "default",
@@ -588,6 +636,7 @@ def delete_old(
 
 
 # TODO: make this return a list of TaskInfo instead?
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def get_tasks(
     num_tasks: int = None, order: Literal["new", "old"] = "new", folder: str = "default"
 ) -> List[Dict]:
@@ -618,6 +667,7 @@ def get_tasks(
     return [task.dict() for task in tasks]
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def estimate_cost(task_id: str) -> float:
     """Compute the maximum FlexCredit charge for a given task.
 
@@ -660,6 +710,7 @@ def estimate_cost(task_id: str) -> float:
     return flex_unit
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def real_cost(task_id: str) -> float:
     """Get the billed cost for given task after it has been run.
 
@@ -678,6 +729,7 @@ def real_cost(task_id: str) -> float:
     return flex_unit
 
 
+@wait_for_connection(wait_time_sec=CONNECTION_RETRY_TIME)
 def test() -> None:
     """Confirm whether Tidy3D authentication is configured. Raises exception if not."""
     try: