+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `web.test()` to simply test if the authentication is configured correctly and raise exception otherwise.
 - Add `SimulationTask.get_running_tasks()` to get a list of running tasks from the server.
 - Argument `scale` to `ModeSolver.plot_field` to control plot scaling.
+- Retry for set number of seconds in web functions if internet connection error.
 
 ### Changed
 - Perfect electric conductors (PECs) are now modeled as high-conductivity media in both the frontend and backend mode solvers, and their presence triggers the use of a preconditioner to improve numerical stability and robustness. Consequently, the mode solver provides more accurate eigenvectors and field distributions when PEC structures are present.